added stronger semantics around unlink if file exists

+++ lib/ramble/ramble/workspace/workspace.py
@@ -1067,8 +1067,10 @@ class Workspace(object):
         out_file = os.path.join(self.root, filename_base + file_extension)
         latest_file = os.path.join(self.root, latest_base + file_extension)
 
-        if overwrite:
-            os.unlink(latest_file) # this is actually removing
+        if overwrite and os.path.exists(latest_file):
+            # this is actually an alias remove, so will delete it regardless of
+            # if it's a simlink or file
+            os.unlink(latest_file)
 
         os.symlink(out_file, latest_file)
